% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_period_station.R
\name{meteo_period_station}
\alias{meteo_period_station}
\title{Returns a tibble with hourly weather data for a period of several days}
\usage{
meteo_period_station(date_beginning, date_end, station_name, station_id)
}
\arguments{
\item{date_beginning}{the date of the beginning of the period of interest, formatted as a year-month-day string (e.g. 2019-03-27)}

\item{date_end}{the date of the end of the period of interest, formatted as a year-month-day string (e.g. 2019-03-27)}

\item{station_name}{the name of the station of interest, for instance "dole-tavaux"}

\item{identifiant_station}{the ID of the station of interest, for instance "07386"}
}
\value{
a tibble
}
\description{
Returns a tibble with hourly weather data for a period of several days
}
\examples{
library(scrapInfoclimat)
meteo_period_station(date_beginning="2018-06-01",
                      date_end="2018-06-03",
                      station_name="dole-tavaux",
                      station_id="07386")
}
