% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_date_station.R
\name{url_date_station}
\alias{url_date_station}
\title{Returns the url that will be scraped}
\usage{
url_date_station(date_ymd, station_name, station_id)
}
\arguments{
\item{station_name}{the name of the station of interest, for instance "dole-tavaux"}

\item{date}{the date of interest, formatted as a year-month-day string (e.g. 2019-03-27)}

\item{identifiant_station}{the ID of the station of interest, for instance "07386"}
}
\value{
a tibble
}
\description{
Returns the url that will be scraped
}
\examples{
library(scrapInfoclimat)
url_date_station(date="2018-06-01",
                 station_name="dole-tavaux",
                 station_id="07386")
}
