% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_date_station.R
\name{weather_date_station}
\alias{weather_date_station}
\title{Returns a tibble a tibble with hourly weather data for one day}
\usage{
weather_date_station(date_ymd, station_name, station_id)
}
\arguments{
\item{station_name}{the name of the station of interest, for instance "dole-tavaux"}

\item{date}{the date of interest, formatted as a year-month-day string (e.g. 2019-03-27)}

\item{identifiant_station}{the ID of the station of interest, for instance "07386"}
}
\value{
a tibble
}
\description{
Returns a tibble a tibble with hourly weather data for one day
}
\examples{
library(scrapInfoclimat)
weather_date_station(date_ymd="2018-06-05",
                   station_name="dole-tavaux",
                   station_id="07386")
}
